[ feat ] 게시판 기능 컴포넌트 리팩토링
1. 서버요청 로직, 값 처리 로직, 인터페이스를 각 커스텀 컴포넌트로 분리하여 재사용성, 유지보수가 용이하게 리팩토링함
