+++ next-researchlab/src/app/strategylab/editPost/[postNo]/page.tsx
@@ -1,24 +1,23 @@
 "use client";
 
 import PostForm from "@/components/strategyLab/postForm";
-import { Member, RootState } from "@/redux/memberActions";
-import { Post } from "@/types/types";
-import axios from "axios";
+import { RootState } from "@/redux/memberActions";
+import postService from "@/services/postService";
+import { Member, Post } from "@/types/types";
 import { useParams } from "next/navigation";
 import { useEffect, useState } from "react";
 import { useSelector } from "react-redux";
 
 const EditPost = () => {
   const params = useParams();
-  const postNo = params?.postNo;
+  const postNo = typeof params?.postNo === "string" ? params.postNo : "";
 
+  const [post, setPost] = useState<Post | null>(null);
   const [isLoading, setIsLoading] = useState(true);
 
   const member = useSelector(
     (state: RootState) => state.member
   ) as Member | null;
-
-  const [post, setPost] = useState<Post | null>(null);
 
   //포스트 정보 서버에 요청
   useEffect(() => {
@@ -26,8 +25,8 @@ const EditPost = () => {
       console.log("상세글 조회");
 
       try {
-        const response = await axios.get(`/api/strategylab/posts/${postNo}`);
-        setPost(response.data);
+        const postDetail = await postService.fetchPostDetail(postNo);
+        setPost(postDetail);
         setIsLoading(false);
       } catch (error) {
         console.log(error);
